move MenuRecordDTOs to seperate package in DTO
